Integrado a funcionalidade de remover uma obra. Otimizado a estrutura do código fonte no front end e removido o campo de contato (no lugar ficou o campo sobre o time de desenvolvimento).

+++ src/main/java/control/recomendation/profile/JSON.java
@@ -23,7 +23,7 @@ public class JSON {
 	    	
 	    	if(obraId.equals(aux)){
 	    		Obra obra = new Obra();
-	    		obra.setid(Integer.valueOf(result.getString("id")));
+	    		obra.setid(Long.valueOf(result.getString("id")));
 	    		obra.settitulo(result.getString("titulo"));
 	    		obra.setresumo(result.getString("resumo"));
 	    		obra.setarea(result.getString("Area"));
